update dependence package
