+++ FE-NextJS/.hintrc
@@ -4,7 +4,6 @@
   ],
   "hints": {
     "typescript-config/strict": "off",
-    "no-inline-styles": "off",
     "typescript-config/consistent-casing": "off"
   }
 }
