+++ FE-NextJS/package.json
@@ -16,34 +16,38 @@
     "deploy:ecosystem": "pm2 start ecosystem.config.cjs"
   },
   "dependencies": {
+    "@atlaskit/pragmatic-drag-and-drop": "^1.4.0",
+    "@atlaskit/pragmatic-drag-and-drop-auto-scroll": "^2.1.0",
+    "@atlaskit/pragmatic-drag-and-drop-hitbox": "^1.0.3",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@hello-pangea/dnd": "^17.0.0",
-    "@mui/icons-material": "^6.3.0",
-    "@mui/material": "^6.3.0",
-    "@mui/styles": "^6.3.0",
-    "@mui/system": "^6.3.0",
-    "@mui/x-date-pickers": "^7.23.3",
-    "@reduxjs/toolkit": "^2.3.0",
+    "@mui/icons-material": "^6.4.0",
+    "@mui/material": "^6.4.0",
+    "@mui/styles": "^6.4.0",
+    "@mui/system": "^6.4.0",
+    "@mui/x-date-pickers": "^7.24.0",
+    "@reduxjs/toolkit": "^2.5.0",
     "@stomp/stompjs": "^7.0.0",
-    "ag-grid-community": "^33.0.3",
-    "ag-grid-react": "^33.0.3",
+    "ag-grid-community": "^33.0.4",
+    "ag-grid-react": "^33.0.4",
     "axios": "^1.7.7",
     "buffer": "^6.0.3",
     "cookie": "^1.0.1",
     "dayjs": "^1.11.13",
     "i18next": "^24.2.1",
     "lodash": "^4.17.21",
-    "next": "^15.1.4",
+    "next": "^15.1.5",
     "node-forge": "^1.3.1",
     "nth-check": "^2.1.1",
     "punycode": "^2.3.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-dropzone": "^14.2.10",
+    "react-dropzone": "^14.3.5",
     "react-hook-form": "^7.53.1",
     "react-i18next": "^15.0.3",
-    "react-redux": "^9.1.2",
+    "react-redux": "^9.2.0",
+    "react-window": "^1.8.11",
     "redux": "^5.0.1",
     "redux-persist": "^6.0.0",
     "request-ip": "^3.3.0",
@@ -54,25 +58,27 @@
     "@eslint/compat": "^1.2.5",
     "@eslint/eslintrc": "^3.2.0",
     "@eslint/js": "^9.18.0",
-    "@next/eslint-plugin-next": "^15.1.4",
+    "@next/eslint-plugin-next": "^15.1.5",
     "@types/lodash": "^4.17.12",
-    "@types/node": "^22.7.7",
+    "@types/node": "^22.10.7",
     "@types/node-forge": "^1.3.11",
-    "@types/react": "^19.0.5",
+    "@types/react": "^19.0.7",
     "@types/react-dom": "^19.0.3",
     "@types/request-ip": "^0.0.41",
     "@types/sockjs-client": "^1.5.4",
-    "@typescript-eslint/eslint-plugin": "^8.19.1",
-    "@typescript-eslint/parser": "^8.19.1",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
+    "css-minimizer-webpack-plugin": "^7.0.0",
     "eslint": "^9.18.0",
-    "eslint-config-next": "^15.1.4",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-config-next": "^15.1.5",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-prettier": "^5.2.3",
     "globals": "^15.14.0",
     "prettier": "^3.3.3",
     "terser-webpack-plugin": "^5.3.11",
     "typescript": "^5.7.3",
-    "typescript-eslint": "^8.19.1"
+    "typescript-eslint": "^8.20.0",
+    "webpack": "^5.97.1"
   },
   "browserslist": {
     "production": [
