+++ FE-NextJS/src/app/(main)/kanban-v2/page.tsx
@@ -1,0 +1,11 @@
+'use server';
+import BackdropLoading from '@/components/loading/BackdropLoading';
+import dynamic from 'next/dynamic';
+
+const KanbanUI = dynamic(() => import('./ui'), { ssr: true, loading: () => <BackdropLoading /> });
+
+const KanbanPage = () => {
+  return <KanbanUI />;
+};
+
+export default KanbanPage;
