+++ FE-NextJS/src/app/(main)/kanban-v2/ui.tsx
@@ -1,0 +1,8 @@
+'use client';
+import { memo } from 'react';
+
+const Home = () => {
+  return <>Kanban v2</>;
+};
+
+export default memo(Home);
